refactor(compose): extract ordering and merge logic from mod.rs
compose/mod.rs had grown to 561 lines, over the 500-line hard limit. Extract
two concerns into their own files, each carrying its tests: order.rs (the
resolve_order / resolve_levels topological ordering) and merge.rs (the YAML
alias-expansion guard and merge-key application). resolve_order/resolve_levels
are re-exported so compose::resolve_order and the crate-root re-export are
unchanged. Pure relocation, no behavior or public-API change.
